requirements.txt
fastapi==0.115.0 
uvicorn==0.23.0 
torch==2.0.1 t
orchvision==0.15.2 
Pillow==9.5.0 
matplotlib==3.7.1